+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## Unreleased
+
+### Added
+
+- Manage secrets using `jdlib.secrets` module
+- `ShellCommandMixin` for executing shell commands
+
 ## [0.0.6] - 2026-03-20
 
 ### Added